+++ NEWS.md
@@ -6,3 +6,14 @@
 
 * rename functions to set and show cache path to avoid conflicts with other packages
 * add functionality for additional tables
+
+# cmhc v0.2.2
+
+* fix table sanfu for Scss by intended market
+* add vignette for rental universe
+
+# cmhc v0.2.3
+
+* fix more table sanfu for Scss by intended market
+* add query builder helper function
+* add access to census and core housing need tables